+++ euc05-sitio-solicitudes-inversiones/euc05-service/src/main/java/com/citi/euces/solicitudes/infra/dto/AutoCetesResponseDTO.java
@@ -8,12 +8,12 @@ public class AutoCetesResponseDTO implements Serializable {
 	
 
 	private Long Id_Plazo;
-	private Long cete;
+	private double cete;
 	/**
 	 * @param id_Plazo
 	 * @param cete
 	 */
-	public AutoCetesResponseDTO(Long id_Plazo, Long cete) {
+	public AutoCetesResponseDTO(Long id_Plazo, double cete) {
 		super();
 		Id_Plazo = id_Plazo;
 		this.cete = cete;
@@ -24,10 +24,10 @@ public class AutoCetesResponseDTO implements Serializable {
 	public void setId_Plazo(Long id_Plazo) {
 		Id_Plazo = id_Plazo;
 	}
-	public Long getCete() {
+	public double getCete() {
 		return cete;
 	}
-	public void setCete(Long cete) {
+	public void setCete(double cete) {
 		this.cete = cete;
 	}
