+++ euc05-sitio-solicitudes-inversiones/euc05-service/src/main/java/com/citi/euces/solicitudes/infra/dto/AutoTasaResponseDTO.java
@@ -8,7 +8,7 @@ public class AutoTasaResponseDTO implements Serializable {
 
     private static final long serialVersionUID = 1L;
 
-	private Long ID_TASAUTO;
+	private BigInteger ID_TASAUTO;
 	private String FECHA_SOLIC;
 	private String FECHA_ESTATUS;
 	private String FECHA_PROCESS;
@@ -56,7 +56,7 @@ public class AutoTasaResponseDTO implements Serializable {
 	 * @param sOEID_OPE
 	 * @param iS_PROCESS
 	 */
-	public AutoTasaResponseDTO(Long iD_TASAUTO, String fECHA_SOLIC, String fECHA_ESTATUS,
+	public AutoTasaResponseDTO(BigInteger iD_TASAUTO, String fECHA_SOLIC, String fECHA_ESTATUS,
 			String fECHA_PROCESS, String eSTATUS, Long sUC_SOLIC, String dIVISION, BigInteger nUM_CTE, String nOM_CTE,
 			BigInteger cONTRATO, String nOMINA, String nOMEJEC, Long mONTO, Long pLAZO, Long tASA_AUTORI,
 			Long nUM_AUTORI_UEC, String oBSERVA_WEB, String jUSTIFICACION, String tIPO_AUTORI, String sOEID_ASIG,
@@ -86,10 +86,10 @@ public class AutoTasaResponseDTO implements Serializable {
 		SOEID_OPE = sOEID_OPE;
 		IS_PROCESS = iS_PROCESS;
 	}
-	public Long getID_TASAUTO() {
+	public BigInteger getID_TASAUTO() {
 		return ID_TASAUTO;
 	}
-	public void setID_TASAUTO(Long iD_TASAUTO) {
+	public void setID_TASAUTO(BigInteger iD_TASAUTO) {
 		ID_TASAUTO = iD_TASAUTO;
 	}
 	public String getFECHA_SOLIC() {
