+++ euc05-sitio-solicitudes-inversiones/euc05-service/src/main/java/com/citi/euces/solicitudes/repositories/SucursalesRepo.java
@@ -2,6 +2,7 @@ package com.citi.euces.solicitudes.repositories;
 
 
 
+import java.math.BigInteger;
 import java.util.List;
 
 import org.springframework.data.jpa.repository.JpaRepository;
@@ -12,7 +13,7 @@ import com.citi.euces.solicitudes.entities.Sucursales;
 
 public interface SucursalesRepo extends JpaRepository<Sucursales, Integer>{
 	@Query(value = "select s.DISTRITO from UEC_CAT_SUC2021 s where s.SIRH =:SIRH",nativeQuery = true )
-	String ObtenerRegionDeNumSuc(@Param("SIRH") Long SIRH);
+	String ObtenerRegionDeNumSuc(@Param("SIRH") BigInteger SIRH);
 	
 	@Query(value = "select * from UEC_CAT_SUC2021 a order by a.SIRH",nativeQuery = true )
 	List<Sucursales> ObtenerSucursalesPorSuc();
